A lot of stuff...
I forgot to committ regularly... But basically, the google api script
was created. There is a pin on the map for every festival I have been
to. Each one has a mouse over that tells you which festival it was and
the years I attended it. And finally, when you click on a pin for a
given festival, it loads the website for that festival..